requirements.txt
python-ldap
python-ldap-test
nose
coverage

